+++ pages/index.vue
@@ -1,8 +1,23 @@
+<script>
+export default {
+  head () {
+    return {
+      bodyAttrs: {
+        class: 'has-background-primary  has-content-centered'
+      }
+    };
+  }
+};
+</script>
+
+
+
+
+
 <template>
-  <img
-    alt="peerhaven logo"
-    src="~/assets/img/ph-circle-darkgrey-white-transparent-200.png"
+  <base-image
+    fallback="png"
+    filename="ph-circle-darkgrey-white-transparent"
+    info="Peerhaven logo"
   />
 </template>
-
-<script src="./index.js"></script>