+++ CHANGELOG.md
@@ -4,10 +4,23 @@
 >
 > (Source: [Keep a Changelog](https://keepachangelog.com/en/1.0.0/))
 
-## 2018-04-25
+## 1.0.1 · 2021-04-27
 
 ### Added
 
-- peerhaven logo
+- Console warning to help prevent [social engineering](https://en.wikipedia.org/wiki/Social_engineering_%28security%29) attacks
+- [Content Security Policy](https://content-security-policy.com/)
+- Error pages
+- License: [Fair Source 5 (v0.9)](https://fair.io/)
+
+### Removed
+
+- Analytics by [Matomo](https://matomo.org/)
+
+## 1.0.0 · 2018-04-25
+
+### Added
+
+- Peerhaven logo
 - [Favicons](https://en.wikipedia.org/wiki/Favicon) for various platforms (iOS, Android, Windows, browsers)
 - Analytics by [Matomo](https://matomo.org/)